src/requirements.txt
Flask>=2.2.0
mysql-connector-python==8.0.33
